fix(cli): correct handling of optional secrets
